slackbot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
dataclasses-json==0.5.7
deprecation==2.1.0
Flask==2.2.2
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.150
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.0
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
python-box==7.0.0
PyYAML==6.0
requests==2.29.0
rsa==4.9
six==1.16.0
slack-bolt==1.16.1
slack-sdk==3.21.2
SQLAlchemy==2.0.11
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.3.0
yarl==1.9.2
